# api/openapi.yaml
openapi: 3.0.3
info:
  title: 智能选股系统 API
  description: 基于Go语言开发的智能选股系统API文档
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.stock.com/v1
    description: 生产环境
paths:
...
  /stocks:
    get:
      summary: 获取股票列表
      tags: [股票管理]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: market
          in: query
          schema:
            type: string
            enum: [SZ, SH, BJ]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockListResponse'
  /stocks/{code}:
    get:
      summary: 获取股票详情
      tags: [股票管理]
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockDetailResponse'

  # 分析相关接口
...
components:
  schemas:
...
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        timestamp:
          type: string
          format: date-time

    # 股票相关模型
    Stock:
      type: object
      properties:
        code:
          type: string
          example: "000001.SZ"
        name:
          type: string
          example: "平安银行"
        market:
          type: string
          example: "SZ"
        industry:
          type: string
          example: "银行"
        list_date:
          type: string
          format: date
        is_active:
          type: boolean
    StockListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                stocks:
                  type: array
                  items:
                    $ref: '#/components/schemas/Stock'
                total:
                  type: integer
                page:
                  type: integer
                limit:
                  type: integer

    # 技术分析模型
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...